Notice interface 모듈화
