+++ src/models/interface/system/Notice/INotice.tsx
@@ -1,0 +1,7 @@
+export interface INotice {
+    notiSeq: number;
+    notiTitle: string;
+    loginId: string;
+    notiContent: string;
+    notiDate: string;
+}
